# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/brokers/summaries": {
            "get": {
                "description": "Find broker summaries by stock code, start date, end date, investor type, and transaction type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Broker"
                ],
                "summary": "Find broker summaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock code",
                        "name": "stock_code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (yyyy-mm-dd)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (yyyy-mm-dd)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Investor type (all, domestic, foreign)",
                        "name": "investor_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction type (all, net, buy, sell)",
                        "name": "transaction_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BrokerSummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/stock": {
            "get": {
                "description": "Find stock by stock code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "Find stock by stock code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stock code",
                        "name": "stock_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StockResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/stocks": {
            "get": {
                "description": "List all stocks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stock"
                ],
                "summary": "List all stocks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.StockResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AuditCommittee": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
...
        "model.BrokerSummaryData": {
            "type": "object",
            "properties": {
                "avg": {
                    "type": "number"
                },
                "broker_code": {
                    "type": "string"
                },
                "lot": {
                    "type": "number"
                },
                "val": {
                    "type": "string"
                }
            }
        },
        "model.BrokerSummaryResponse": {
            "type": "object",
            "properties": {
                "buyers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BrokerSummaryData"
                    }
                },
                "end_date": {
                    "type": "string"
                },
                "sellers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BrokerSummaryData"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "stock_code": {
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/model.Summary"
                }
            }
        },
        "model.Commissioner": {
            "type": "object",
            "properties": {
                "is_independent": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "model.Director": {
            "type": "object",
            "properties": {
                "is_affiliated": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
        "model.Dividend": {
            "type": "object",
            "properties": {
                "cash_dividend_currency": {
                    "type": "string"
                },
                "cash_dividend_per_share": {
                    "type": "number"
                },
                "cash_dividend_per_share_currency": {
                    "type": "string"
                },
                "cash_dividend_total": {
                    "type": "number"
                },
                "cum_date": {
                    "type": "string"
                },
                "ex_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "payment_date": {
                    "type": "string"
                },
                "ratio_1": {
                    "type": "integer"
                },
                "ratio_2": {
                    "type": "integer"
                },
                "record_date": {
                    "type": "string"
                },
                "total_stock_bonus": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "year": {
                    "type": "string"
                }
            }
        },
...
        "model.Profile": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bae": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "listing_date": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "main_business": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "stock_code": {
                    "type": "string"
                },
                "stock_name": {
                    "type": "string"
                },
                "sub_industry": {
                    "type": "string"
                },
                "sub_sector": {
                    "type": "string"
                },
                "tin": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "model.Secretary": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fax": {
                    "type": "string"
                },
                "mobile_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "model.Shareholder": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "is_controller": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                },
                "share": {
                    "type": "number"
                }
            }
        },
        "model.StockResponse": {
            "type": "object",
            "properties": {
                "audit_committees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AuditCommittee"
                    }
                },
                "board": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "commissioners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Commissioner"
                    }
                },
                "directors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Director"
                    }
                },
                "dividends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Dividend"
                    }
                },
                "listing_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "profiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Profile"
                    }
                },
                "secretaries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Secretary"
                    }
                },
                "share": {
                    "type": "number"
                },
                "shareholders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Shareholder"
                    }
                },
                "subsidiaries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Subsidiary"
                    }
                }
            }
        },
...
        "model.Subsidiary": {
            "type": "object",
            "properties": {
                "business_fields": {
                    "type": "string"
                },
                "commercial_year": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operation_status": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                },
                "total_asset": {
                    "type": "number"
                },
                "units": {
                    "type": "string"
                }
            }
        },
        "model.Summary": {
            "type": "object",
            "properties": {
                "avg": {
                    "type": "number"
                },
                "foreign_net_val": {
                    "type": "string"
                },
                "total_lot": {
                    "type": "number"
                },
                "total_val": {
                    "type": "string"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}